kyles branch for test

+++ imx291_driver/imx291/src/i2c.h
@@ -16,7 +16,7 @@ typedef struct addr_val {
 int delay_ms(unsigned int msec);
 void sensors_ADC_init(uint16_t addr);  
 void write8(uint16_t regAddr, uint8_t value); 
-uint16_t read8(uint16_t regAddr, uint8_t dev_addr);
+uint8_t read8(uint16_t regAddr, uint8_t dev_addr);
 int delay_ms(unsigned int msec);
 int i2c_writeArr(AddrVal *regStructArr);
